ocr-text-detection-backend/requirements.txt
azure-core 
azure-ai-formrecognizer
python-multipart